merging probe detect with main

+++ drivers/sensor/afe4400/afe4400.c
@@ -63,13 +63,13 @@ static int afe4400_sample_fetch(const struct device *dev, enum sensor_channel ch
     uint32_t led1val = _afe4400_read_reg(dev, LED1VAL);
     led1val = (uint32_t) (led1val << 10);
     int32_t led1val_signed = (int32_t) led1val;
-    drv_data->raw_sample_ir = led1val_signed>>10;
+    drv_data->raw_sample_ir = (int32_t) led1val_signed>>10;
 
     _afe4400_reg_write(dev, CONTROL0, 0x000001);
     uint32_t led2val = _afe4400_read_reg(dev, LED2VAL);
     led2val = (uint32_t) (led2val << 10);
     int32_t led2val_signed = (int32_t) led2val;
-    drv_data->raw_sample_red = led2val_signed>>10;
+    drv_data->raw_sample_red = (int32_t) led2val_signed>>10;
 
     return 0;
 }
